Renamed variables related to the 3rd column of every bunker segment
Also split code into different functions so each function won't be so
long. But this may increase the work that the machine has to do since
images are loaded after the entire grid is created, instead of being
loaded along with the grid; which means the computer has to traverse the
grid twice?