WIP: fire on each spacebar press
Each bullet has its own interval timer. Multiple bullets can be
displayed at a time. If need be, I can try using the
same interval for all bullets; this may be required if multiple
intervals cause the game to lag.

The original game allows only one bullet to displayed at a time, so the
player can't spam bullets to win and has to be strategic about when and
where to shoot. I'll try that in the next commit.